Refactor: Modularize SongDetailScreen and SongListScreen
This commit refactors `SongDetailScreen.tsx` and `SongListScreen.tsx` to improve readability, maintainability, and separation of concerns. No functional changes have been made.

Key changes include:

1.  **`SongListScreen.tsx` Refactoring:**
    *   Extracted `SongListItem.tsx`: A new component to render each individual song item in the list. `SongListScreen.tsx` now uses this component within its `FlatList`.
    *   Relevant styles were moved to `SongListItem.tsx`.

2.  **`SongDetailScreen.tsx` Refactoring:**
    *   Extracted `SongDisplay.tsx`: A new component responsible for rendering the song's HTML content within a `WebView`. It handles its own loading state display.
    *   Extracted `useSongProcessor.ts`: A custom hook that encapsulates all logic for parsing ChordPro content, applying formatting (transposition, font changes, chord visibility), and generating the final HTML for the `SongDisplay`. It manages the `songHtml` and processing state.
    *   Extracted `SongControls.tsx`: A new component that contains all user interface controls for the song detail view, including Floating Action Buttons (FABs) for actions like toggling chords, transposing, changing font size/family, and their associated Modals. This component manages its internal state for modal visibility.
    *   `SongDetailScreen.tsx` is now significantly leaner. Its primary roles are fetching initial song data, managing core song display settings (e.g., `currentTranspose`, `chordsVisible`), and coordinating the `SongDisplay`, `useSongProcessor`, and `SongControls`.
    *   Styles were moved to their respective new components (`SongDisplay.tsx`, `SongControls.tsx`).

This modularization makes each part of the songbook functionality easier to understand, and modify independently.
